fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
strawberry-graphql